+++ internal/controller/vintagemc_controller.go
@@ -20,53 +20,66 @@ import (
 	"context"
 	"fmt"
 
-	"github.com/giantswarm/logging-operator/pkg/capicluster"
 	loggingreconciler "github.com/giantswarm/logging-operator/pkg/logging-reconciler"
+	"github.com/giantswarm/logging-operator/pkg/vintagemc"
 	"github.com/pkg/errors"
+	corev1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/runtime"
 	"k8s.io/apimachinery/pkg/types"
-	capiv1beta1 "sigs.k8s.io/cluster-api/api/v1beta1"
 	ctrl "sigs.k8s.io/controller-runtime"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 	"sigs.k8s.io/controller-runtime/pkg/log"
 )
 
-// ClusterReconciler reconciles a Cluster object
-type ClusterReconciler struct {
+// VintageMCReconciler reconciles a Service object
+type VintageMCReconciler struct {
 	client.Client
 	Scheme            *runtime.Scheme
 	LoggingReconciler loggingreconciler.LoggingReconciler
 }
 
-//+kubebuilder:rbac:groups=cluster.x-k8s.io,resources=clusters,verbs=get;list;watch
-//+kubebuilder:rbac:groups=cluster.x-k8s.io,resources=clusters/status,verbs=get
+//+kubebuilder:rbac:groups=core,resources=services,verbs=get;list;watch
+//+kubebuilder:rbac:groups=core,resources=services/status,verbs=get
 
 // Reconcile is part of the main kubernetes reconciliation loop which aims to
 // move the current state of the cluster closer to the desired state.
-// It compares the state specified by the Cluster object against the actual
-// cluster state, and then perform operations to make the cluster state reflect
-// the desired state.
+// TODO(user): Modify the Reconcile function to compare the state specified by
+// the Service object against the actual cluster state, and then
+// perform operations to make the cluster state reflect the state specified by
+// the user.
 //
 // For more details, check Reconcile and its Result here:
 // - https://pkg.go.dev/sigs.k8s.io/controller-runtime@v0.14.4/pkg/reconcile
-func (r *ClusterReconciler) Reconcile(ctx context.Context, req ctrl.Request) (result ctrl.Result, err error) {
+func (r *VintageMCReconciler) Reconcile(ctx context.Context, req ctrl.Request) (result ctrl.Result, err error) {
 	logger := log.FromContext(ctx)
 
-	cluster := &capiv1beta1.Cluster{}
-	err = r.Client.Get(ctx, types.NamespacedName{Name: req.Name, Namespace: req.Namespace}, cluster)
+	// We only want to watch 1 service, kubernetes in default NS.
+	if req.Namespace != "default" || req.Name != "kubernetes" {
+		return ctrl.Result{}, nil
+	}
+
+	service := &corev1.Service{}
+	err = r.Client.Get(ctx, types.NamespacedName{Name: req.Name, Namespace: req.Namespace}, service)
 	if err != nil {
 		// TODO(theo): might need to ignore when objects are not found since we cannot do anything
 		//             see https://book.kubebuilder.io/reference/using-finalizers.html
 		//if r.Client.IsNotFound(err) {
-		//	return ctrl.Result{}, nil
+		//  return ctrl.Result{}, nil
 		//}
 		return ctrl.Result{}, errors.WithStack(err)
 	}
 
-	logger.Info(fmt.Sprintf("Name %s", cluster.GetName()))
+	logger.Info(fmt.Sprintf("Reconciling Management cluster"))
 
-	loggedCluster := capicluster.Object{
-		Object: cluster,
+	// TODO(theo): Pass the IsLoggingEnabled function as a parameter into the LoggingReconciler
+	// So we can have different detection logic to enable logging for Vintage MC and CAPI cluster.
+	// On Vintage MC we determine if logging is enabled based on a global installation
+	// level setting which need to be passed as a flag to this operator.
+	// IsLoggingEnabled for this controller would only check the given flag, while the CAPI controller
+	// would check both the flag and the label.
+
+	loggedCluster := vintagemc.Object{
+		Object: service,
 	}
 	result, err = r.LoggingReconciler.Reconcile(ctx, loggedCluster)
 	if err != nil {
@@ -77,8 +90,8 @@ func (r *ClusterReconciler) Reconcile(ctx context.Context, req ctrl.Request) (re
 }
 
 // SetupWithManager sets up the controller with the Manager.
-func (r *ClusterReconciler) SetupWithManager(mgr ctrl.Manager) error {
+func (r *VintageMCReconciler) SetupWithManager(mgr ctrl.Manager) error {
 	return ctrl.NewControllerManagedBy(mgr).
-		For(&capiv1beta1.Cluster{}).
+		For(&corev1.Service{}).
 		Complete(r)
 }