+++ pkg/vintagewc/cluster.go
@@ -24,3 +24,7 @@ func (o Object) GetAppsNamespace() string {
 func (o Object) AppConfigName(app string) string {
 	return app
 }
+
+func (o Object) GetObject() client.Object {
+	return o.Object
+}